Fix segfault caused by uncleared stack
The scilib module was causing a segfault when the program looped arround without clearing the stack.
This occurred because the stack was not properly cleared before reusing it, leading to invalid memory access.
To address this issue, a new function, `clearStack`, was added to the stack module.
The `clearStack` function ensures that the stack is emptied before reusing it, preventing any segfaults caused by accessing invalid memory.

[stack]
Add the clear stack function in stack.c, and it's interface in stack.h.

[scilib]
Add a reset function, that resets the value of the index of buffer (postind) to zero, and calls the clearStack() function.